requirements.txt
requests>=2.25.0
requests_toolbelt>=0.9.1
gql>=3.4.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.0
pandas>=1.3.0
pydantic>=1.8.0
python-dotenv>=0.19.0
tenacity>=8.0.0
click>=8.0.0
rich>=12.0.0
loguru>=0.6.0
pytest>=7.0.0
black>=22.0.0
mypy>=0.950
